refactor(cli): extract the tick-pass wiring from the entry point
Move the observe → plan → act helper and its Worker/PR/conflict-Worker
collaborator wiring out of cli.ts into src/tick.ts, leaving the entry
point with only usage text, flag parsing/validation, command routing,
and exit-code assignment. Pure prefactor ahead of the stricli
migration (#27); no operator-visible change.

Closes #28
